refactor(esphome): split bridge component into Coordinator + Bridge + Worker
Foundation for Mode B (auto-discovery pair flow). No user-visible
behavior change in Mode A — same services, same events, same
lazy-encrypt + auth-backoff semantics.

- ShaverCoordinator (new, plain C++): GATT logic, state machine,
  lazy-encrypt probe (8d560100/8d560117), auth-failure backoff,
  subscription handling. Mode-agnostic — receives raw GAP/GATT
  events through callbacks from a Worker.
- ShaverBridge (new, Component + CustomAPIDevice): HA service
  registration (ble_read_char, ble_subscribe, ble_unsubscribe,
  ble_write_char, ble_set_throttle, ble_get_info), heartbeat loop,
  connected sensor, event emission. Owns no GATT state.
- PhilipsShaver (trimmed): BLEClientNode adapter that forwards
  events to the Coordinator and registers a set_enabled callback
  (the wrapper exposes set_enabled(), BLEClientBase doesn't).
- __init__.py: instantiates Coordinator + Bridge + Worker, wires
  them, injects shared log_tag. Schema unchanged in Mode A.

Compile-validated against ESP-IDF on m5stack-atom (esphome 2026.4.5).
Hardware smoke (XP9201 + OneBlade reconnect, AtomS3R reboot Issue #6
regression) pending.
